adds embedded figure binaries in json upload
Adds initial support for uploading obojobo module .json files
that can include image binary data in the json.  Previously it
was not possible to import a json file with the images referenced
in it's chunks due to them referencing ID's in specific obojobo
instance they were being created on.

This change allows figure chunks to support a new content property
named 'imageBinary' that contains base64 encoded image data when
uploading via the dashboard in .json formatted modules.

This also provides a script that can be run against an existing
.json module on your local files and populate imageBinary for you.

It will will attempt to locate images by their content.filename
in the directory (or sub directory) of the .json module you provide

See embedFigureBinary.js for usage.

This change does not support embedding image binaries in xml formatted
modules and I have not tested support for uploading .json in the editor.

(cherry picked from commit 9cc4a4f7401709417807b5543effc031900ec315)

+++ packages/app/obojobo-express/server/models/media.js
@@ -376,7 +376,13 @@ class Media {
 		let file
 
 		try {
-			file = await fs.readFile(fileInfo.path)
+			if (fileInfo.buffer) {
+				// allow fileInfo to already contain a loaded buffer
+				file = fileInfo.buffer
+			} else {
+				// load the file from disk
+				file = await fs.readFile(fileInfo.path)
+			}
 		} catch (error) {
 			// calling methods expect a thenable object to be returned
 			logger.logError('Error Reading media file', error)
@@ -387,7 +393,7 @@ class Media {
 
 		if (!isValid) {
 			// Delete the temporary media stored by Multer
-			await fs.unlink(fileInfo.path)
+			if (fileInfo.path) await fs.unlink(fileInfo.path)
 			throw new Error(
 				`File upload only supports the following filetypes: ${mediaConfig.allowedMimeTypesRegex
 					.split('|')
@@ -408,7 +414,7 @@ class Media {
 			})
 
 			// Delete the temporary media stored by Multer
-			await fs.unlink(fileInfo.path)
+			if (fileInfo.path) await fs.unlink(fileInfo.path)
 
 			oboEvents.emit(Media.EVENT_IMAGE_CREATED, {
 				userId,
